msm: mdss: Remove pwm_disable from backlight pwm driver
The PWM implementation has been reconfigured so that pwm_enable only
needs to be called once for the life of the MDSS driver. The new
implementation allows the driver to call pwm_config in order to change
the backlight without needing to disable and enable. The backlight PWM
driver is disabled when the panel is turned off, and reenabled when the
panel comes up.

Change-Id: Ieba05921b53eb1cf686aaa2751f9cc3650522bb7
Signed-off-by: Benet Clark <benetc@codeaurora.org>