ASoC: msm: Fix the noise in HDMI on HDMI connect disconnect
Pops and long squeak sounds are heard during the resolution switch
on some monitors
The noise is due to the playback of the stale data in the end device.
To workaround the issue, a silence has to be played all the times to
clear out the end device buffers.
Open a playback session with the HDMI device, this will cause a underrun
in the AFE port whenver the no audio playback is in progress.
This underrun will trigger silence packets at the AFE port and will
clear out the end device stale buffers

Change-Id: Idcc3a417747b8692fe73fbf682662678e97af64e
Signed-off-by: Banajit Goswami <bgoswami@codeaurora.org>