msm: mdss: Fix the issue of recovery from DCM state
In the QDCM tool, toggling from DCM state to DTM state fails
because system un-blank doesn't happen successfully. The root
cause of the issue is at the DCM_BLANK, the overlay_off is called
before setting the panel_power_state to OFF, due to which
overlay_off doesn't suspend properly. Therefore, rearrange the
order to fix the problem.

Change-Id: I130fe711dc0d679d8205d22a9027d1c179ed49d5
Signed-off-by: Zohaib Alam <zalam@codeaurora.org>